+++ domain/src/main/java/com/healthcare/management/system/domain/DomainUserDAO.java
@@ -1,5 +1,7 @@
 package com.healthcare.management.system.domain;
 
+import com.healthcare.management.system.domain.model.DomainUser;
+
 import java.util.List;
 
 public interface DomainUserDAO {
