+++ domain/src/main/java/com/healthcare/management/system/domain/model/DomainUser.java
@@ -1,4 +1,4 @@
-package com.healthcare.management.system.domain;
+package com.healthcare.management.system.domain.model;
 
 import lombok.Data;
